Decouple Graph display from database display and change max row in table
